[fix] 게시판 delete 수정
[fix] 게시판 delete 수정
what : 게시판 글 삭제시 멤버 판별코드 추가
why : 게시판 글 삭제시 멤버 판별코드가 존재 안했음

+++ Backend/src/main/java/com/github/backend/service/post/impl/PostServiceImpl.java
@@ -63,9 +63,17 @@ public class PostServiceImpl implements PostService {
     }
 
     @Override
-    public PostInfoOutPutDto.Info deletePost(Long postId) {
+    public PostInfoOutPutDto.Info deletePost(Long postId, MemberInfo memberInfo) {
+        Member member = memberRepository.findById(memberInfo.getId())
+                .orElseThrow(() -> new PostException(PostErrorCode.MEMBER_NOT_EXISTS));
+
         Post post = postRepository.findById(postId)
                 .orElseThrow(() -> new PostException(PostErrorCode.POST_NOT_EXISTS));
+
+        if(!Objects.equals(member.getId(), post.getMember().getId())){
+            throw new PostException(PostErrorCode.MEMBER_NOT_EXISTS);
+        }
+
         postRepository.delete(post);
 
         return PostInfoOutPutDto.Info.of((post));
